Separacion de responsabilidades, ahora app.jsx se enfoca en la lógica principal y el header, mientras VideoCarousel maneja toda la interacción con las cards
